+++ tsdoc/src/__tests__/ParsingBasics.test.ts
@@ -1,5 +1,5 @@
 import {
-  CoreModifierTagSet,
+  StandardModifierTagSet,
   DocComment,
   ParserContext
 } from '../index';
@@ -16,11 +16,11 @@ test('01 Simple @beta and @internal extraction', () => {
   ].join('\n'));
 
   const docComment: DocComment = parserContext.docComment;
-  const modifierTagSet: CoreModifierTagSet = docComment.modifierTagSet;
+  const modifierTagSet: StandardModifierTagSet = docComment.modifierTagSet;
 
-  expect(modifierTagSet.isAlpha).toEqual(false);
-  expect(modifierTagSet.isBeta).toEqual(true);
-  expect(modifierTagSet.isInternal).toEqual(true);
+  expect(modifierTagSet.isAlpha()).toEqual(false);
+  expect(modifierTagSet.isBeta()).toEqual(true);
+  expect(modifierTagSet.isInternal()).toEqual(true);
 });
 
 test('02 A basic TSDoc comment with all components', () => {